Refactor statushandler into CRUD (#2858)
### Changed

- Refactor statushandler into CRUD
