+++ cg/store/api/core.py
@@ -1,6 +1,5 @@
 import logging
 
-from cg.store.api.status import StatusHandler
 from cg.store.crud.create import CreateHandler
 from cg.store.crud.delete import DeleteDataHandler
 from cg.store.crud.read import ReadHandler
@@ -14,7 +13,6 @@ class CoreHandler(
     CreateHandler,
     DeleteDataHandler,
     ReadHandler,
-    StatusHandler,
     UpdateHandler,
 ):
     """Aggregating class for the store api handlers."""
@@ -22,7 +20,6 @@ class CoreHandler(
     def __init__(self, session):
         DeleteDataHandler(session)
         ReadHandler(session)
-        StatusHandler(session)
         UpdateHandler(session)
