+++ tests/store/api/conftest.py
@@ -10,12 +10,6 @@ from cg.store.models import CaseSample
 from tests.store_helpers import StoreHelpers
 
 
-@pytest.fixture(name="max_nr_of_cases")
-def max_nr_of_cases() -> int:
-    """Return the number of maximum number of cases"""
-    return 50
-
-
 @pytest.fixture(name="store_failing_sequencing_qc")
 def store_failing_sequencing_qc(
     bcl2fastq_flow_cell_id: str,
@@ -61,56 +55,6 @@ def store_failing_sequencing_qc(
         sample_base_percentage_passing_q30=30,
     )
     return store
-
-
-@pytest.fixture(name="max_nr_of_samples")
-def max_nr_of_samples() -> int:
-    """Return the number of maximum number of samples"""
-    return 50
-
-
-@pytest.fixture(name="three_customer_ids")
-def three_customer_ids() -> list[str]:
-    """Return three customer ids."""
-    yield ["".join(["cust00", str(number)]) for number in range(3)]
-
-
-@pytest.fixture(name="three_pool_names")
-def three_pool_names() -> list[str]:
-    """Return three customer ids."""
-    yield ["_".join(["test_pool", str(number)]) for number in range(3)]
-
-
-@pytest.fixture(name="store_with_samples_for_multiple_customers")
-def store_with_samples_for_multiple_customers(
-    store: Store, helpers: StoreHelpers, timestamp_now: dt.datetime
-) -> Store:
-    """Return a store with two samples for three different customers."""
-    for number in range(3):
-        helpers.add_sample(
-            store=store,
-            customer_id="".join(["cust00", str(number)]),
-            internal_id="_".join(["test_sample", str(number)]),
-            no_invoice=False,
-            delivered_at=timestamp_now,
-        )
-    yield store
-
-
-@pytest.fixture(name="store_with_pools_for_multiple_customers")
-def store_with_pools_for_multiple_customers(
-    store: Store, helpers: StoreHelpers, timestamp_now: dt.datetime
-) -> Store:
-    """Return a store with two samples for three different customers."""
-    for number in range(3):
-        helpers.ensure_pool(
-            store=store,
-            name="_".join(["test_pool", str(number)]),
-            customer_id="".join(["cust00", str(number)]),
-            no_invoice=False,
-            delivered_at=timestamp_now,
-        )
-    yield store
 
 
 @pytest.fixture(name="store_with_analyses_for_cases")
