+++ tests/store/crud/read/test_read_customer.py
